Fixes for polygon editing
